Finished refactoring postprocess.py so that the plotter and cruncher are seperate classes each inheriting from a base processor class. Added some command line arguments to postprocess to either only crunch or only plot, which I think will come in handy. Also finished the convergence plot.
TODO:

1. Add logging to stdout and a log file for all scripts. Each script gets
its own log and they should be put in a special subdirectory of the basedir
for the run
2. Optimize array manipulations. Everything runs really slowly and I suspect
methods like column_stack and array_split are the culprit. These methods can
be replaced with some clever indexing/slicing which should be much faster. This
might required a seperate timing script as well.
3. Add a line plot that plots a specified quantity along a given path in 3D space