Updated TODO list
Refactored postprocess by adding a Simulation object. This object contains
all the information about a simulation as well as its data, but does't not
peform any processing on that data. It is mostly a convenient container that
knows how to manage its data (read and write, extract columns, etc). Now instead
of setting a sim as a attribute in the processor classes, this simulation object
is passed around between the relevant processing functions. This makes parellization
possible and also just makes way more sense in general.

Note! We need to be careful about clearing all the data from a simulation object when
we are done with it so we don't consume tons of memory.