Added adaptive convergence to lua simulation script. Converted everything to use a Simulator object and broke out the various components of setting up and running a simulation into different instances methods.
Added local angular averaging of a given quantity to postprocessing

Added global averaging of local averages to postprocessing

Added plotting of global averages. Completely assumes its an angular
average of some sort, so its a little rough right now. Needs to be
generalized and whatnot.

Fixed comsol_compare so it always has the normE columns in the correct
place