Added code to compute transmission, reflection, and absorption. Pretty sure this is the correct way to do it but might need to be double checked. Can check absorbance against the other method of integrating the squared magnitude of the electric field to ensure accuracy.
Added code to compute Jsc. Right now the answer it returns is horribly wrong. I imagine
its a unit problem. The result for the generation rate is also wrong, differing from Anna's
by a factor of 10^4. This is again a unit problem. I think I may be computing a volumetric gen
rate, and not an areal generation rate (or something like that). Need to investigate further.