Modified postprocess a bit. The notion of "global averages" was kind of silly and didn't really fit into what I was trying to compute. I've now added a scalar_reduce function that reduces some scalar quantity on a grid (normE, normEsquared, genRate) across all simulations. You can either compute a direct sum (default, as in when computing total generation) or you can pass True to the avg kwarg to compute an average.
Also added super limited geometry drawing for 2d heatmaps and slightly adjusted
the heatmap2d call signature to only require a sim object if you are drawing
some geometry. It might be possible to remove this dependency on a sim object
in the future to make the function more general. The drawing basically only
works for my nanowire system right now, until I deal with the geometry of
handling a general shape.

In the future, I think it will be really beneficial to organize all the data in
an SQLite database. It will really prevent all the awkwardness of having data
organized in a relatively static tree of directories. It will add the ability
to perform more complicated queries, and will also provide a single, unified
interface to the data. Right now dealing will diverse files, with their unique
column order/heading structure and formats, makes peforming the same actions on
two different pieces of data non-uniform.

I also think the postprocessing library could use a make over. I'm just not
completely satisfied with the interface it provides. It feels awkward, and
there is also redundant loading of data from different simulations. A
SimulationGroup object might help, but overall I think the answer is to stick
all the data in an SQLite database.