Some minor modifications to the postprocessing script to fix ylimits and expand environment variables in the configs and such
Added a destructor method to the Simulator object to clean up the file
descriptor used by the log handlers. There was a bug where these wouldn't get
cleaned up and for large job campaigns would hit the file descriptor limit of
the OS

Added a check in sim_wrapper to make sure the combined core shell diameter
doesn't exceed the array period. For now just skips the simulation if it does
but it would make much more sense to have a prevalidation function that gets
called on all the simulation configs before even entering the map_async
function.

The rest should just be hacks on the plotting scripts for PVSC