More docs
Removed superfluous log and print messages

Added ability to provide and optional argument to transmissionData
function of Cruncher class to specify location of transmission port.
This makes it so we don't have to modify the code when running
optimizations/postprocessing to see how location of the transmission
port affects things.