Added functionality to sim_wrapper.py and simulator.py to write all data and simulation configs to and HDF5 file. Lightly tested and seems to be working well so far. Saving to directories containing npz and text files is still possible
Quick notes about implementation:

Pytables doesn't support concurrent writes. I implemented a FileWriter class
that inherits from threading.Thread. This thread manages all writing to the
HDF5 file. It has a queue stored as an attribute and its run() method
continuously polls this queue.. The subprocesses that are actually running a
simulation push a tuple onto this queue when they wish to write something to
the file. The tuple contains the method of the FileWriter class to call, a
tuple of the arguments to pass to the function, and a dict of kwargs to pass to
the function in that order.

Caveats:

When using HDF5, log files are still saved to the filesystem under directories
labeled by the sim id. There are various ways to deal with this. We could save
the log files as a big string in an attribute of the parent group (just like
the sim configs). The other option would be to use the filenode subpackage
of PyTables to store a "file-like" node under the sim group. I'm tabling that
decision for now so I can make this commit.