Lazy-loading Feature Addition
This is a huge change to the postprocessing code that I'm super excited
about. Data is now "lazy-loaded" within the .data attribute of each
instance of a Simulation object. This means the possibility to access
the data is always available, but the data is only pulled from this when
you attempt to access it.

This is accomplished via a DataManager object that inherits from
collections.MutableMapping. The DataManager behaves likes a dict, but
upon instantiation updates it's keys with all available data items in
the on-disk data file, storing None as the corresponding value. When a
key is accessed, if it's value is None the __getitem__ method calls a
_load_data method to pull the data from disk. The data then stays in the
dict until destruction.

The DataManager is also responsible for all the writing of data to the
disk. Methods of the Simulation object for reading/writing data are
basically just wrappers around the methods of its DataManager instance.

This also enables a really nicely pluggable architecture for different
storage backends. Just write a new DataManager subclass for your storage
backend, then add logic to the _get_data_manager factory function to
instantiate the correct manager based on the save_as entry in the config
file.

A quick test seems to indicate everything is working. There are a few
things I would like to test/add

1. Make sure KeyErrors are properly thrown if the user attempts to
access a nonsense key
2. Use some sort of hashing check to determine if data has changed at
all. If it hasn't changed, we can skip writing it to disk and save a
bunch of time.