Lazy Write Feature
Added functionality to only update data in the data dict and write it
out to disk if it has changed in some way. This works great when using
the HDF5 backend, but doesn't work for NPZ files because I can find a
way of reaching in and modifying individual arrays in the archive. This
could be solved by writing my own wrapper around a collection of
individual npy files, but this is way more trouble than its worth. Just
use HDF5.