Postprocessing redesign
Introduced the SimulationGroup object. This is an object representing a
group of simulations, and have methods fo calculating things that only
make sense for a group of simulations.

Totally redesigned the postprocessing script. Now all the methods
defined in the config file are methods of either a Simulation or a
SimulationGroup object. All the local methods are methods of a
Simulation object, and all the global methods are methods of a
SimulationGroup object. Now only a single Processor object exists, and
this object just calls methods of the relevant Simulation or
SimulationGroup objects. This cut down on a lot of repeated code. Its
also slightly more generic, so we could define any kind of "process" in
the config file. This process would be represented by an object, and
would have methods that could be specified under the "process" heading
in the config file.