Parallelized Postprocessing
Parallelized the postprocessing code. This required a few somewhat hacky
fixes that I'm not entirely comfortable with, but it works.

Basically, the Simulation and SimulationGroup objects are not
pickleable, so we can't pass them to child processes. Thus, I had to do
a few things:

1) Move all the crunching/plotting class methods into module level
functions
2) Move instantiation of the relevant objects into these module level
functions. I pass the path to the config files into these functions
instead of the actual objects, and instantiate from the paths.
3) Make a replace class method on Processor that takes the self.sims and
self.sim_groups attributes of the Processor object and turns the objects
therein into strings containing the path to their config files. Strings
are pickleable so we can hand this to pool.map
4) Make the crunching/plotting class methods into simple wrappers around
the corresponding module level functions

The whole replace thing is kind of icky and makes me uncomfortable and
kind of ruins the interface in my opinion. I might be able to make a
nice way to do this behind the scenes so object only get replaced with
strings right before we pass them off to pool.map