Incident Power Bug Fix
Fixed an edge case when computing incident power for a single frequency.
It's unphysical to claim there is power contained in a truly
monochromatic incident wave. If we want to simulate at a single
frequency, then we just default to using the power of the NREL bin that
contains the desired frequency. I set the bin_size to 0 then have some
logic to handle that in _get_incident_amplitude.

Added some warnings about trying to plot complex quantities in
simulation.py