Major Reorganization 2
Moved everything in simulator.py out of utils and into simulate.py

Moved simulation.py out of utils and into optics directory

Moved the DataManager classes out of simulation.py and into their own
file

Added a run_optics script for running a job campaign, added this script
to the setup.py file

Majorly improved the logging in simulate.py. It's a bit messy right now,
as I've moved the logic into the top of the module. But, now if an
exception gets raised in a child process of the Pool object, the ID of
that simulation and the traceback get logged to the console, and logged
to the main simulate.log file. The traceback also gets logged to that
simulations log file.