Fourier Coefficients
Added code to extract the eigenmode amplitudes of a solution. This isn't
actually as useful as it sounds, because the physical meaning of the
mode amplitudes for patterned layers is extremely unclear. For unpatterned
layers, one needs to use the GetDiffractedOrders function to retrieve
the 1-based index needed to extract the mode amplitude corresponding to
a particular point in k-space. This function is not exposed in the
Python API, and I may need to implement it myself.