New Postprocessing Function
Added a function to compute the absorption in a layer using two
different methods and compare them. One integrates the field intensity,
and one uses the power flux at the top and bottom of a layer

Just writes to a text file for now. TODO: Will need to fix the data
saving to write to the HDF5 file like its supposed to.