Config Object Updates, Postprocesing Additions
Fixed interpolation-before-evaluation errors in the config object.
Things seem to be working well at this point, still no unit tests
though.

Hacking on layer absorption in postprocessing and trying to get the
volume integral of the intensity to agree with the difference between
the poynting flux at the top and bottom of each layer.

Moved make_logger calls in simulate.py to _after_ the directory for
the simulation has been created