Major Config Overhaul
This is a major, breaking change to the config file format that is not
compatible with any older simulation config files.

Optimizable and variable parameters can now be located anywhere in the
config, rather than having to be located in special "params" sections
(which was super dumb). This just required adding a recursive
_find_params function to the Config object. Parameters are identified by
the 'paramtype' key located immediately below the parameter name.
Example below:

    param_name:
      paramtype:
      iterype:
      start:
      end:
      step:
      value:

paramtype must specify the type of parameter (currently only 'variable'
or 'optimized'). Itertype is only used if paramtype = 'variable'.
Currently itertype must be one of ['list', 'numsteps', 'stepsize'].
If 'list', value must be populated with a list of explicit values.
If 'stepsize' or 'numsteps', start, end, and step must be populated with
the appropriate values.

When individual sim_conf files are written, variable and optimized
parameters no longer contain dicts but instead just contain unnested,
raw values. This eliminates the need for the conf['param_name']['value']
syntax which was dumb.

The rest of the changes are just updates to handle this new
specification.