Geometry Code Changes
Moved the get_layers() method from the Simulation object in
postprocess.py into a module level function in geometry.py. This just
makes way more sense because both Simulator and Simulation objects can
utilize the functionality. You just pass in the Simulator/Simulation
object as a argument and get back an OrderedDict of parameters.

Also added the InPlaneSamplingConvergence notebook, which is not working
and correctly demonstrates why convergence of the absorbed power is bad
with respect to the number of in plane sampling points. Its because of
the step discontinuities introduced by material parameters in layers
that contain patterning.