Added ability to update z sampling
Added ability to update sampling along the z direction without
overwriting old samples. This merges the arrays from the two z sampling
schemes and stores the result in the HDF5 file.

Required adding flag to CLI in run_optics

Added some utility functions for merging and sorting in utils.py

Added core functionality to simulate.py

Fixed minor things in postprocess.py

Fixed a bug in data_manager that tried to save array with the same key
but different shape into the corresponding node in the HDF5 file. In
this case, we need to delete the HDF5 node and recreate with the new
shape.