Added ability to specify list of z sampling points
The user can not specify the number of z sampling points, or explicitly
pass in a list of z values they want to take samples at. Specifying the
number of z sampling points results in linearly spaced points between 0
and max_depth.

This is useful for non-uniform sampling along the z direction, with a
perhaps increased density near the interfaces between layers.