Feature Additions
Added an integrate_layer function to Simulation in postprocess.py that
allows integrating nkEsq in a layer by material. It constructs an
linear interpolator from the data points for each material so one can
experiment with the density of points. Currently a bit hacky and not
really ready for production. Just added for experimentation/testing
purposes

Added a fast cartesian_product function for numpy array to utils.py