Major Improvements to Postprocessing Code, NOTE: API CHANGES
Major improvements to the postprocessing code for running postprocessing
calculations in parallel. Instead of storing a list of
Simulation/SimulationGroup objects in the Processor object, I now just
store a list of Config objects instead of a list of Simulations, and a
list of lists of Config objects for SimulationGroups. This is useful
because Config objects are serializable, and can be passed to child
processes using the multiprocessing Pool object.

NOTE: This introduced a major change in the specification of the
Postprocessing sections in the input config file. These still need to be
properly documented.

Added a filtering CLI interface to the run_optics script and simulate.py
that allows running only a subset of the parameters specified in the
master config file by filtering out any unwanted parameters.

TODO: Maybe think about pulling out filter_by, group_against, and
group_by into utils.py

Switched the Simulator object to store a DataManager object in its .data
attribute and use that for writing data to disk instead of handling all
its own writing. This reduced duplication of code and honestly should
have been done a long time ago.

Added some new docs. These were all autogenerated.