Cleanups and Bug Fixes
Cleaned up sample_dict implementation in simulate.py. This requires
sample_dict by present in the General section of the config file. It can
be set to False (or something else False-y) if you do not want to
collect fields by layer.

Fixed up the plotting code in postprocess.py so layer boundaries and
internal geometry get drawn correctly

Enhanced the inheritance scheme in data_manager.py with the use of
Abstract Base Classes. Certain methods are now required to be overridden
by child classes, or an error will be thrown.