Finalized Dispy Functionality. CONFIG FILE CHANGES
Introduced the Dispy library for submitting jobs in parallel both to the
local host from which you are submitting jobs, and remote hosts you have
access to. See the docs for how to properly set this up.

This functionality requires a 'nodes' entry and a 'ip_addr' entry in the
General section of the config file.

The 'nodes' entry specifies the nodes you want to include in your
cluster, and must be a list whose elements may be any of the following
things

- Strings, which must be IP addresses in dotted decimal format.
- One exception is the special 'local' string. This will lookup the address of
  the local host and use that
- length-two lists containing [str, int] = [ip_address, port_number].

The 'ip_addr' entry tells Dispy the IP address of the local host so
remote nodes can send job results back to the host. Must be a string
containing the IP address of the local host, or the string 'auto', in
which case the public facing ip address of the local host will be
automatically discovered and used.

Added some suporting functions in utils.py to handle grabbing IP
addresses and whatnot.