Bug Fixes and Plotting Updates
Added ability to plot the propagation constants q both for a single
simulations and for a a group of simulations when they are grouped
against number of basis terms.

Fixed bugs with passing the grouped_by and grouped_against parameters
along to the SimulationGroup objects.

Fixed bugs with passing the skip_hash kwargs through when trying to re
run a single simulation.