S4 Bug Fix and Path Bug Fixes
Added commit which fixes critical bug in S4.

Fixed some path bugs in simulate.py, data_manager.py, etc

Added a new ID field to the General section of the config for storing
the full SHA hash of the config, instead of just the first 10 characters
stored in sim_dir.

Added a --print flag to process_optics to just print the simulations
that would be processed and exit. Useful with --filter_by for
determining which directories contain sims with a given set of
parameters