Read Configs from Table and Thickness Sweeps
Added the ability to read nested configs which are stored in a PyTables
table. Required writing some supporting functions in utils.py and a
classmethod to Config that reads from a structured numpy array.

Changed the self.sim_confs data structure in SimulationManager to a dict
whose keys are sim IDs and whose values are a tuples of (Config,
conf_path) pairs.

Reorganized preprocess submodule a bit so its easier to import.

Added proper handling of thickness sweeps. Strategy is to group against
thicknesses, then pick only one simulation in each group to run. After
all those run, hard links are made from the directory of the other sims
in the thickness sweep to the solution file of the sim that ran. Then
the remaining sims run and are able to load the solution from disk
instead of recomputing.