Added Support for Quantity Objects
Added support for Quantity objects in the YAML representation of a
config. Required adding a simple function to the Preprocessor and a
simple constructor/representor to the yaml loader/dumper.

Also required slight reworking of the simulations database, because it
is troublesome trying to store arbitrary Python objects in PyTables
Tables. Instead of trying to do that, only the magnitude of the Quantity
is stored in the table. The entire YAML string representation of the
config is also stored in a column, and whenever we want to load the
Config from the table we just use that string instead of trying to
loaded it by recursing through the nested columns in the row.