Bug Fixes and Enhancements
The yaml  code in the config.py was not using the C-based yaml loader/dumper,
causing poor performance. This has been fixed.

Added a dump_configs command to the optics CLI for writing configs stored
in the HDF5 database to YAML files on disk.

Modified the default filters in HDF5DataManager to use zlib compression lib
with compression level of 8