Performance improvement to dump_configs
Realized there is zero need to parse the yaml column in the HDF5
database, because we don't need the resulting data structure. We can
just dump the raw yaml directly to disk. DERP!

There was also a dump_configs function in config.py that I forgot about.
I just enhanced that a little bit and used that