Working UnQLite Datastore!
Got everything working! It required a slightly modified approach.

Configs are still prepared and stored in a document collection. However,
rather than storing configs solely in a document collection and
attempting to deserialize by recursing through the dict returned from
the collection, configs are stored directly in the key-value store in
the database. The key is the Config ID, and the value is the raw bytes
resulting from pickling the Config object. To retrieve a config object,
just pickle.loads the byte stream.

An additional key-value pair for the Config object is stored. The key is
the Config ID with the string '_docid' appended to it. The value is an
integer ID which is the document ID of the document corresponding to the
Config stored in the document collection.

This setup is really useful. One can query the document collection to
find configs corresponding to a set of conditions, then get the Config
ID from the resultings documents and load then from the key value store
by ID. If one knows the ID ahead of time, one can us that directly to
update the bytes in the key-value store, then also grab the document ID
and use that to update the document in the document collection.