Improved Rescaling Method & Modified Query Syntax
Moved the rescaling technique into a get hook which gets applied to the
DataManager object. It had to be done this way because otherwise the
field arrays of all Simulations in a SimulationGroup accumulated in
memory and took up an unreasonable amount of space.

Modified the Query syntax so other elements in a DB document can be
accessed. Also now required the leading / to represent root of the
document. To get at Simulation params use /_d/path/to/param

Added a script for managing the database. Needs some cleaning up,
enhancements, and installation code.