Added Support for Nonuniform Frequency Sweeps
Added support to postprocessing for nonuniform frequency sweeps. It just
looks left and right for the halfway point between the "center"
frequency for the bin and the nearest center frequencies on either side
and stores them in a tuple that gets passed to Simulation objects.
Center is slighty misleading because the representative frequency for a
bin could be off center depending on how nonuniform the sweep is. This
required modifications to function in utils for calculating incident
power of a bin.

Fixed bug in field rescaling hook that removed units attached to an
array.