Bug Fixes and Performance Improvements
Added a get_shape function to the DataManager class that allows
retrieving the shape of a node in an HDF5 file without reading data into
memory. Using this method greatly improves the performance of and makes
usable the _validate_existing_data function, which before read
everything into memory which usually ends up consuming all available
memory and is super slow.

Fixed major bug with indexing/slicing in the averaging procedure of the
angular averaging code.